requirements.txt
Flask
waitress
behave
beautifulsoup4
selenium
allure-behave